Rename ymd dates utils file
