+++ apps/client-react/src/views/project-route.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Link, useNavigate, useParams } from "react-router-dom";
 
 import type { Project } from "@projectsbuild/types";
-import { ymdPretty } from "~/utils/ymd-dates";
+import { ymdPretty } from "~/utils/date-ymd";
 import { useProjectsContext } from "./projects-route";
 
 import styles from "./project-route.module.css";
